requirements.txt
dotenv==0.9.9
  python-dotenv==1.1.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.2
  blinker==1.9.0
  click==8.1.8
  itsdangerous==2.2.0
  Jinja2==3.1.6
    MarkupSafe==3.0.2
  MarkupSafe==3.0.2
  Werkzeug==3.1.3
    MarkupSafe==3.0.2
goose3==3.1.19
  requests==2.32.5
    charset-normalizer==3.4.3
    idna==3.10
    urllib3==2.5.0
    certifi==2025.8.3
  pillow==11.3.0
  lxml==6.0.2
  cssselect==1.3.0
  beautifulsoup4==4.13.5
    soupsieve==2.8
    typing_extensions==4.15.0
  python-dateutil==2.9.0.post0
    six==1.17.0
  langdetect==1.0.9
    six==1.17.0
  pyahocorasick==2.2.0
gTTS==2.5.4
  requests==2.32.5
    charset-normalizer==3.4.3
    idna==3.10
    urllib3==2.5.0
    certifi==2025.8.3
  click==8.1.8
nltk==3.9.1
  click==8.1.8
  joblib==1.5.2
  regex==2025.9.18
  tqdm==4.67.1
pipdeptree==2.28.0
  packaging==25.0
  pip==25.1.1
pt_core_news_md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.8.0/pt_core_news_md-3.8.0-py3-none-any.whl#sha256=54382cda034e41f3ec605ceeb924cd6cfdbced00a65f7afa12218520ba7008c5
pydub==0.25.1
python-slugify==8.0.4
  text-unidecode==1.3
python-telegram-bot==22.5
  httpx==0.28.1
    anyio==4.11.0
      idna==3.10
      sniffio==1.3.1
    certifi==2025.8.3
    httpcore==1.0.9
      certifi==2025.8.3
      h11==0.16.0
    idna==3.10
scikit-learn==1.7.2
  numpy==2.3.3
  scipy==1.16.2
    numpy==2.3.3
  joblib==1.5.2
  threadpoolctl==3.6.0
spacy==3.8.7
  spacy-legacy==3.0.12
  spacy-loggers==1.0.5
  murmurhash==1.0.13
  cymem==2.0.11
  preshed==3.0.10
    cymem==2.0.11
    murmurhash==1.0.13
  thinc==8.3.6
    blis==1.3.0
      numpy==2.3.3
    murmurhash==1.0.13
    cymem==2.0.11
    preshed==3.0.10
      cymem==2.0.11
      murmurhash==1.0.13
    wasabi==1.1.3
    srsly==2.5.1
      catalogue==2.0.10
    catalogue==2.0.10
    confection==0.1.5
      pydantic==2.11.9
        annotated-types==0.7.0
        pydantic_core==2.33.2
          typing_extensions==4.15.0
        typing_extensions==4.15.0
        typing-inspection==0.4.1
          typing_extensions==4.15.0
      srsly==2.5.1
        catalogue==2.0.10
    setuptools==80.9.0
    numpy==2.3.3
    pydantic==2.11.9
      annotated-types==0.7.0
      pydantic_core==2.33.2
        typing_extensions==4.15.0
      typing_extensions==4.15.0
      typing-inspection==0.4.1
        typing_extensions==4.15.0
    packaging==25.0
  wasabi==1.1.3
  srsly==2.5.1
    catalogue==2.0.10
  catalogue==2.0.10
  weasel==0.4.1
    confection==0.1.5
      pydantic==2.11.9
        annotated-types==0.7.0
        pydantic_core==2.33.2
          typing_extensions==4.15.0
        typing_extensions==4.15.0
        typing-inspection==0.4.1
          typing_extensions==4.15.0
      srsly==2.5.1
        catalogue==2.0.10
    packaging==25.0
    wasabi==1.1.3
    srsly==2.5.1
      catalogue==2.0.10
    typer==0.19.1
      click==8.1.8
      typing_extensions==4.15.0
      shellingham==1.5.4
      rich==14.1.0
        markdown-it-py==4.0.0
          mdurl==0.1.2
        Pygments==2.19.2
    cloudpathlib==0.22.0
    smart_open==7.3.1
      wrapt==1.17.3
    requests==2.32.5
      charset-normalizer==3.4.3
      idna==3.10
      urllib3==2.5.0
      certifi==2025.8.3
    pydantic==2.11.9
      annotated-types==0.7.0
      pydantic_core==2.33.2
        typing_extensions==4.15.0
      typing_extensions==4.15.0
      typing-inspection==0.4.1
        typing_extensions==4.15.0
  typer==0.19.1
    click==8.1.8
    typing_extensions==4.15.0
    shellingham==1.5.4
    rich==14.1.0
      markdown-it-py==4.0.0
        mdurl==0.1.2
      Pygments==2.19.2
  tqdm==4.67.1
  numpy==2.3.3
  requests==2.32.5
    charset-normalizer==3.4.3
    idna==3.10
    urllib3==2.5.0
    certifi==2025.8.3
  pydantic==2.11.9
    annotated-types==0.7.0
    pydantic_core==2.33.2
      typing_extensions==4.15.0
    typing_extensions==4.15.0
    typing-inspection==0.4.1
      typing_extensions==4.15.0
  Jinja2==3.1.6
    MarkupSafe==3.0.2
  setuptools==80.9.0
  packaging==25.0
  langcodes==3.5.0
    language_data==1.3.0
      marisa-trie==1.3.1
SpeechRecognition==3.14.3
  typing_extensions==4.15.0
  standard-aifc==3.13.0
    standard-chunk==3.13.0
    audioop-lts==0.2.2
  audioop-lts==0.2.2
telegram==0.0.1
tornado==6.5.2
transformers==4.57.1
  filelock==3.20.0
  huggingface-hub==0.36.0
    filelock==3.20.0
    fsspec==2025.10.0
    packaging==25.0
    PyYAML==6.0.3
    requests==2.32.5
      charset-normalizer==3.4.3
      idna==3.10
      urllib3==2.5.0
      certifi==2025.8.3
    tqdm==4.67.1
    typing_extensions==4.15.0
    hf-xet==1.2.0
  numpy==2.3.3
  packaging==25.0
  PyYAML==6.0.3
  regex==2025.9.18
  requests==2.32.5
    charset-normalizer==3.4.3
    idna==3.10
    urllib3==2.5.0
    certifi==2025.8.3
  tokenizers==0.22.1
    huggingface-hub==0.36.0
      filelock==3.20.0
      fsspec==2025.10.0
      packaging==25.0
      PyYAML==6.0.3
      requests==2.32.5
        charset-normalizer==3.4.3
        idna==3.10
        urllib3==2.5.0
        certifi==2025.8.3
      tqdm==4.67.1
      typing_extensions==4.15.0
      hf-xet==1.2.0
  safetensors==0.6.2
  tqdm==4.67.1
Wikipedia-API==0.8.1
  requests==2.32.5
    charset-normalizer==3.4.3
    idna==3.10
    urllib3==2.5.0
    certifi==2025.8.3
wordcloud==1.9.4
  numpy==2.3.3
  pillow==11.3.0
  matplotlib==3.10.6
    contourpy==1.3.3
      numpy==2.3.3
    cycler==0.12.1
    fonttools==4.60.0
    kiwisolver==1.4.9
    numpy==2.3.3
    packaging==25.0
    pillow==11.3.0
    pyparsing==3.2.5
    python-dateutil==2.9.0.post0
      six==1.17.0
